feat: 입력할 때 빈 문자에 대한 validate 추가, 컴포넌트 하단 분리
